First Real Commit
Added art for player, enemy, ball glove, and bullet.

Player can move around

Ball glove sticks to character and rotates around to where the mouse is
pointing.
The enemy's weapon sticks to the enemy in the same way and points towards
the player when the enemy can see the player.
Player and enemy face to the left when they are aiming left and face to
the right when they are facing right.