New art for the Gun, Player, and Room
Added a gun sprite.
Guns now fade out when they are reloading

Made the character out of a head, body, and backpack so they can be
animated separately.
The backpack opens when the player catches a bullet.
The glove closes and fades a bit when it catches a bullet

Added a hardwood floor and bar counter sprite to create the level.