Butterfly Net, Thug visuals, Rifle, Shotgun
Player can press right click to swing their butterfly net in front of
them. The net can catch as many bullets as it touches. Net swings in a
small arc in front of where the player is aiming.

Bullets now stop moving when they hit something. When they are caught by
the glove, they get stuck to the glove until the glove puts them in the
backpack. The net will carry them in a similar fashion, but doesn't yet.
When bullets hit the player, they will stick there until the player pries
them out to show how much damage the player has taken (but this isn't
implemented yet either).

Tried a few different ways to get the butterfly net to swing in an arc
where the player is aiming. Some remnants are still in the Player script
probably. Like the time it takes to swing and cooldown which aren't being
used at the moment.

Added more sprites for the thugs. They each contain pants and a head, but
can choose between 3 body types, bald, cool hair, a hat, or a bandana.
Added a rifle and shotgun sprite.
Added some wall sprites.