End Screen Tweaks
Bullets now spawn in around the storage area, not all exactly on it.
The bullets now travel in a wider stream, not a straight line.
Bullets start travelling 2 frames apart instead of 1.

Bullets only get added to the array for the first 25. After that, they
still get teleported, but don't get added to the array so they won't move
at the end. They should get destroyed instead of teleported.