Entrance
Added an entrance sprite.
An entrance so you can leave the saloon without dying. Now that I write it
out, that seems pretty minor for a game jam. Just die to see your boss.

Shoudn't run out of space in the array anymore. If the array is full,
destroy bullets. If you caught too many bullets, only iterate through til
the end of the size of the array (whichever is smaller).